backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# Database
psycopg2-binary>=2.9,<3.0
redis>=5.0,<6.0

# Async Processing
celery>=5.3,<6.0
django-celery-beat>=2.5,<3.0
django-celery-results>=2.5,<3.0

# Authentication & Security
djangorestframework-simplejwt>=5.3,<6.0
django-cryptography>=1.1,<2.0
argon2-cffi>=23.1,<24.0

# Multi-tenancy & Auditing
django-simple-history>=3.4,<4.0

# Logging & Monitoring
structlog>=23.2,<24.0
django-prometheus>=2.3,<3.0

# Development & Testing
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
Faker>=20.1,<21.0

# Code Quality
ruff>=0.1,<1.0
black>=23.12,<24.0
mypy>=1.7,<2.0
django-stubs>=4.2,<5.0
djangorestframework-stubs>=3.14,<4.0

# Environment Management
python-decouple>=3.8,<4.0
python-dotenv>=1.0,<2.0

# Validation & Serialization
pydantic>=2.5,<3.0
pydantic-settings>=2.1,<3.0

# Utils
python-dateutil>=2.8,<3.0
phonenumbers>=8.13,<9.0
validate-docbr>=1.10,<2.0

# Production
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0